Suppression des invoices et files

+++ assets/js/pages/MaintenancesPage.jsx
@@ -1,7 +1,7 @@
 import moment from 'moment';
-import React, { useEffect, useState } from 'react';
-import { Link } from 'react-router-dom';
-import { toast } from 'react-toastify';
+import React, {useEffect, useState} from 'react';
+import {Link} from 'react-router-dom';
+import {toast} from 'react-toastify';
 import TableLoader from '../components/loaders/TableLoader';
 import Pagination from '../components/Pagination';
 import MaintenancesAPI from '../services/maintenancesAPI';
@@ -32,15 +32,17 @@ const MaintenancesPage = props => {
 
     //Gestion de la suppression d'une maintenance
     const handleDelete = async id => {
-        const originalMaintenances = [...maintenances];
-        setMaintenances(maintenances.filter(maintenance => maintenance.id !== id));
+        if (confirm("Supprimer ?")) {
+            const originalMaintenances = [...maintenances];
+            setMaintenances(maintenances.filter(maintenance => maintenance.id !== id));
 
-        try {
-            await MaintenancesAPI.delete(id);
-            toast.success("La maintenance a été supprimée !");
-        } catch (error) {
-            toast.error("Une erreur est survenue !");
-            setMaintenances(originalMaintenances);
+            try {
+                await MaintenancesAPI.delete(id);
+                toast.success("La maintenance a été supprimée !");
+            } catch (error) {
+                toast.error("Une erreur est survenue !");
+                setMaintenances(originalMaintenances);
+            }
         }
     };
 
@@ -50,7 +52,7 @@ const MaintenancesPage = props => {
     };
 
     //Gestion de la recherche
-    const handleSearch = ({ currentTarget }) => {
+    const handleSearch = ({currentTarget}) => {
         setSearch(currentTarget.value);
         setCurrentPage(1);
     };
@@ -68,15 +70,15 @@ const MaintenancesPage = props => {
             m.vehicule.brand.toLowerCase().includes(search.toLowerCase()) ||
             m.vehicule.reference.toLowerCase().includes(search.toLowerCase())
     );
-    
+
     //Pagination des maintenances
     const paginatedMaintenances = Pagination.getData(
-        filteredMaintenances, 
-        currentPage, 
+        filteredMaintenances,
+        currentPage,
         itemsPerPage
     );
 
-    return ( 
+    return (
         <>
             <div className="mb-3 d-flex justify-content-between align-items-center">
                 <h1>Liste des maintenances</h1>
@@ -84,69 +86,69 @@ const MaintenancesPage = props => {
             </div>
 
             <div className="form-group">
-                <input 
-                    type="text" 
+                <input
+                    type="text"
                     onChange={handleSearch}
                     value={search}
-                    className="form-control" 
+                    className="form-control"
                     placeholder="Rechercher..."
                 />
             </div>
 
             <table className="table table-hover">
                 <thead>
-                    <tr>
-                        <th>Chrono</th>
-                        <th className="text-center">Date</th>
-                        <th>Type</th>
-                        <th className="text-center">Montant</th>
-                        <th>Véhicule</th>
-                        <th/>
-                    </tr>
+                <tr>
+                    <th>Chrono</th>
+                    <th className="text-center">Date</th>
+                    <th>Type</th>
+                    <th className="text-center">Montant</th>
+                    <th>Véhicule</th>
+                    <th/>
+                </tr>
                 </thead>
                 {!loading && <tbody>
-                    {paginatedMaintenances.map(maintenance =>
-                        <tr key={maintenance.id}>
-                            <td>{maintenance.chrono}</td>
-                            <td className="text-center">{formatDate(maintenance.date)}</td>
-                            <td>{maintenance.type}</td>
-                            <td className="text-center">
-                                {maintenance.amount.toLocaleString()} €
-                            </td>
-                            <td>
-                                {maintenance.vehicule.brand} 
-                                {maintenance.vehicule.reference}
-                            </td>
-                            <td>
-                                <Link 
-                                    to={"/maintenances/" + maintenance.id} 
-                                    className="btn btn-sm btn-primary mr-1">
-                                        Modifier
-                                </Link>
-                                <button 
-                                    onClick={() => handleDelete(maintenance.id)}
-                                    className="btn btn-sm btn-danger">
-                                        Supprimer
-                                </button>
-                            </td>
-                        </tr>
-                    )}
-                </tbody> }
+                {paginatedMaintenances.map(maintenance =>
+                    <tr key={maintenance.id}>
+                        <td>{maintenance.chrono}</td>
+                        <td className="text-center">{formatDate(maintenance.date)}</td>
+                        <td>{maintenance.type}</td>
+                        <td className="text-center">
+                            {maintenance.amount.toLocaleString()} €
+                        </td>
+                        <td>
+                            {maintenance.vehicule.brand}
+                            {maintenance.vehicule.reference}
+                        </td>
+                        <td>
+                            <Link
+                                to={"/maintenances/" + maintenance.id}
+                                className="btn btn-sm btn-primary mr-1">
+                                Modifier
+                            </Link>
+                            <button
+                                onClick={() => handleDelete(maintenance.id)}
+                                className="btn btn-sm btn-danger">
+                                Supprimer
+                            </button>
+                        </td>
+                    </tr>
+                )}
+                </tbody>}
             </table>
 
-            {loading && <TableLoader />}
+            {loading && <TableLoader/>}
 
             {itemsPerPage < filteredMaintenances.length &&
-                <Pagination 
-                    currentPage={currentPage} 
-                    itemsPerPage={itemsPerPage} 
-                    length={filteredMaintenances.length} 
-                    onPageChanged={handlePageChange} 
-                />
+            <Pagination
+                currentPage={currentPage}
+                itemsPerPage={itemsPerPage}
+                length={filteredMaintenances.length}
+                onPageChanged={handlePageChange}
+            />
             }
 
         </>
-     );
+    );
 };
- 
+
 export default MaintenancesPage;
